Share dateFromParts, fix date formatting
